geoserver: wait for DB and DB user to be available
Waiting just for Postgres to be accepting TCP connections is not enough.
The Postgres service is slow to set up during the first deploy. GeoFence
would fail to create its tables (as the DB is not ready yet), and stay
broken instead of crashing or trying again.

Note that an image including psql is necessary.